[datakit] Split source protocol parsers by format
Keep Terminus and OpenCode parsing in separate download modules. Each
source selects the parser and handles its dataset-specific prompt fields.
Remove the terminal_chat module and its shared format dispatcher.
